+++ src/Validator/FormatEnabled.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Mtarld\ApiPlatformMsBundle\Validator;
+
+use Symfony\Component\Validator\Constraint;
+
+/**
+ * @final @internal
+ * @Annotation
+ * @psalm-suppress PropertyNotSetInConstructor
+ */
+class FormatEnabled extends Constraint
+{
+    /**
+     * @var string
+     *
+     * Violation message
+     */
+    public $message = "'{{ format }}' format is not enabled.";
+}