+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.2.0] - Unreleased
+* Loaded serialization configuration files only when specified in a microservice config
 * Moved services definition from XML to PHP
 * Added classes to preload
 * Added `create`, `update` and `delete` methods in `AbstractHttpMicroserviceRepository`